requirements.txt
# 视频处理
opencv-python==4.8.1.78

# OCR文字识别
pytesseract==0.3.10
Pillow==10.0.1

# 数值计算和图像处理
numpy==1.24.3
scikit-image==0.21.0
scipy==1.10.1
imagehash==4.3.1

# 文件系统监控
watchdog==3.0.0
pathlib2==2.3.7

# 命令行界面
click==8.1.7
colorama==0.4.6

# 图形用户界面 (通常Python内置)
# tkinter (内置，无需安装)

# Development dependencies
flake8
mypy